requirements.txt
streamlit==1.38.0
transformers==4.44.2
torch
Pillow
numpy